Skapat metoder i GUIMethods för att koppla upp mot databasen samt läsa in db.properties Tagit bort onödiga metoder från loginGUI och TimeTrackGUI
